Forked aosp's private BulkCursor APIs for forward compatibility.
Instead of linking against the private APIs we now copy the
implementation of them into AndroidSQLiteServer and use the public APIs
that they depend on (save for ContentObserver#getContentObserver).  This
makes it both easier to be backwards compatible and also gives us
important future proofing.

Tested that Gingerbread now works properly.