Fixed a few bugs highlighted by turning on test code coverage.
In particular, we were not calling close() on a lot of tests.  Adding it
triggered some regressions where we were not safe about handling
performRelease on the thread affinity executor.

Also identified a small race during thread "handoff" in ThreadedTest.
This was a bug in the test itself, not in AndroidSQLiteServer.

P.S. Code coverage is being reported as 51%.  And that number is low
because the code coverage tool is not currently capturing code paths
taken by other processes (or so it would seem).